backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI & Agent Framework
google-generativeai==0.3.1
langchain==0.1.0
langchain-google-genai==0.0.5
langgraph==0.0.20

# Twitter Integration
tweepy==4.14.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1


